+++ lib/src/screens/favourite_notes.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:note_it/src/widgets/no_notes_info.dart';
+
+class FavouriteNotesScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Favourite Notes'),
+        centerTitle: true,
+        actions: <Widget>[
+          IconButton(icon: Icon(Icons.search), onPressed: () {}),
+          IconButton(icon: Icon(Icons.more_vert), onPressed: () {}),
+        ],
+      ),
+      body: NoNoteInfo()
+    );
+  }
+}
